refactor: add given fixture test
