+++ api-server/src/test/java/nexters/payout/apiserver/dividend/application/DividendQueryServiceTest.java
@@ -4,50 +4,23 @@ import nexters.payout.apiserver.dividend.application.dto.request.DividendRequest
 import nexters.payout.apiserver.dividend.application.dto.request.TickerShare;
 import nexters.payout.apiserver.dividend.application.dto.response.MonthlyDividendResponse;
 import nexters.payout.apiserver.dividend.application.dto.response.YearlyDividendResponse;
-import nexters.payout.core.time.InstantProvider;
-import nexters.payout.domain.DividendFixture;
-import nexters.payout.domain.StockFixture;
-import nexters.payout.domain.dividend.domain.Dividend;
-import nexters.payout.domain.dividend.domain.repository.DividendRepository;
-import nexters.payout.domain.stock.domain.Sector;
-import nexters.payout.domain.stock.domain.Stock;
-import nexters.payout.domain.stock.domain.repository.StockRepository;
+import nexters.payout.apiserver.dividend.common.GivenFixtureTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
-import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.BDDMockito.given;
 import static nexters.payout.domain.StockFixture.*;
 import static nexters.payout.domain.stock.domain.Sector.*;
 
 @ExtendWith(MockitoExtension.class)
-class DividendQueryServiceTest {
+class DividendQueryServiceTest extends GivenFixtureTest {
 
     @InjectMocks
     private DividendQueryService dividendQueryService;
-
-    @Mock
-    private DividendRepository dividendRepository;
-
-    @Mock
-    private StockRepository stockRepository;
-
-    private final Integer JANUARY = 1;
-    private final Integer DECEMBER = 12;
 
     @Test
     void 사용자의_월간_배당금_정보를_가져온다() {
@@ -98,60 +71,5 @@ class DividendQueryServiceTest {
                 new TickerShare(AAPL, 2),
                 new TickerShare(TSLA, 1),
                 new TickerShare(SBUX, 1)));
-    }
-
-    private void givenStockAndDividendForMonthly(String ticker, Sector sector, double dividend, int... cycle) {
-        Stock stock = StockFixture.createStock(ticker, sector);
-        given(stockRepository.findByTicker(eq(ticker))).willReturn(Optional.of(stock));
-
-        for (int month = JANUARY; month <= DECEMBER; month++) {
-            if (isContain(cycle, month)) {
-                // 배당 주기에 해당하는 경우
-                given(dividendRepository.findAllByTickerAndYearAndMonth(
-                        eq(ticker),
-                        eq(InstantProvider.getLastYear()),
-                        eq(month)))
-                        .willReturn(List.of(DividendFixture.createDividendWithExDividendDate(
-                                stock.getId(),
-                                dividend,
-                                parseDate(InstantProvider.getLastYear(), month)
-                        )));
-            } else {
-                // 배당 주기에 해당하지 않는 경우
-                given(dividendRepository.findAllByTickerAndYearAndMonth(
-                        eq(ticker),
-                        eq(InstantProvider.getLastYear()),
-                        eq(month)))
-                        .willReturn(new ArrayList<>());
-            }
-        }
-    }
-
-    private void givenStockAndDividendForYearly(String ticker, Sector sector, double dividend, int... cycle) {
-        Stock stock = StockFixture.createStock(ticker, sector);
-        given(stockRepository.findByTicker(eq(ticker))).willReturn(Optional.of(stock));
-
-        List<Dividend> dividends = new ArrayList<>();
-        for (int month : cycle) {
-            dividends.add(DividendFixture.createDividendWithExDividendDate(
-                    stock.getId(),
-                    dividend,
-                    parseDate(InstantProvider.getLastYear(), month)));
-        }
-
-        given(dividendRepository.findAllByTickerAndYear(
-                eq(ticker),
-                eq(InstantProvider.getLastYear())))
-                .willReturn(dividends);
-    }
-
-    private boolean isContain(int[] cycle, int month) {
-        return Arrays.stream(cycle).anyMatch(m -> m == month);
-    }
-
-    private Instant parseDate(int year, int month) {
-        LocalDate date = LocalDate.of(year, month, 1);
-        ZonedDateTime zonedDateTime = date.atStartOfDay(ZoneId.of("UTC"));
-        return zonedDateTime.toInstant();
     }
 }
